Fixes constant restarting hopefully.

+++ Sistine/monitors/swearFilter.js
@@ -12,6 +12,7 @@ module.exports = class swearFilter extends Monitor {
 	}
 
 	async run(msg) {
+		if (msg.channel.type !== 'text' && !msg.guild.available) return null;
 		const { configs } = msg.guild;
 		if (!configs.filtering.swearFilter || configs.filtering.swearWords.length >= 0) return null;
 		if (msg.deletable && (new RegExp(configs.filtering.swearWords.join('|'), 'i')).test(msg.content)) await msg.delete();
